requirements.txt
telebot
pyTelegramBotAPI
py-cpuinfo
logzero
zipfile36
speedtest-cli
freeport==0.1.9
psutil
docker